requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
MarkupSafe==2.1.2
packaging==23.0
prompt-toolkit==3.0.38
PyJWT==2.6.0
pyTelegramBotAPI==4.10.0
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
router==0.1
ruamel.yaml==0.17.21
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
